Add Tree performance tests
