+++ CollectionsPerformanceTests/TreePerformanceTests.swift
@@ -1,0 +1,14 @@
+//
+//  TreePerformanceTests.swift
+//  Collections
+//
+//  Created by James Bean on 7/11/17.
+//
+//
+
+import XCTest
+import Collections
+
+class TreePerformanceTests: XCTestCase {
+
+}
